@@ -18,7 +18,7 @@
     "config": {
         "sort-packages": true,
         "platform": {
-            "php": "7.3.99"
+            "php": "8.1.99"
         },
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
@@ -31,15 +31,15 @@
         }
     },
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0",
+        "php": "~8.1.0 || ~8.2.0 || ~8.3.0",
         "laminas/laminas-stdlib": "^3.7.1"
     },
     "require-dev": {
-        "laminas/laminas-coding-standard": "~2.2.1",
-        "laminas/laminas-eventmanager": "^3.4.0",
-        "laminas/laminas-hydrator": "^3.2 || ^4.3",
-        "laminas/laminas-servicemanager": "^3.7.0",
-        "phpunit/phpunit": "^9.5.19"
+        "laminas/laminas-coding-standard": "^2.4.0",
+        "laminas/laminas-eventmanager": "^3.6.0",
+        "laminas/laminas-hydrator": "^4.7",
+        "laminas/laminas-servicemanager": "^3.19.0",
+        "phpunit/phpunit": "^9.5.25"
     },
     "suggest": {
         "laminas/laminas-eventmanager": "Laminas\\EventManager component",
