@@ -18,7 +18,7 @@
     "config": {
         "sort-packages": true,
         "platform": {
-            "php": "7.4.99"
+            "php": "8.1.99"
         },
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
@@ -27,15 +27,14 @@
     "extra": {
     },
     "require": {
-        "php": "^7.4 || ~8.0.0 || ~8.1.0"
+        "php": "~8.1.0 || ~8.2.0 || ~8.3.0"
     },
     "require-dev": {
-        "laminas/laminas-coding-standard": "~2.3.0",
-        "phpbench/phpbench": "^1.2.6",
-        "phpunit/phpunit": "^9.5.23",
-        "psalm/plugin-phpunit": "^0.17.0",
-        "vimeo/psalm": "^4.26",
-        "phpstan/phpdoc-parser": "^0.5.4"
+        "laminas/laminas-coding-standard": "^2.5",
+        "phpbench/phpbench": "^1.2.15",
+        "phpunit/phpunit": "^10.5.8",
+        "psalm/plugin-phpunit": "^0.18.4",
+        "vimeo/psalm": "^5.20.0"
     },
     "autoload": {
         "psr-4": {
