@@ -12,7 +12,6 @@ use UnexpectedValueException;
 
 use function array_map;
 use function count;
-use function get_class;
 use function is_array;
 use function serialize;
 use function sprintf;
@@ -96,10 +95,9 @@ class PriorityQueue implements Countable, IteratorAggregate, Serializable
      * the same item has been added multiple times, it will not remove other
      * instances.
      *
-     * @param  mixed $datum
      * @return bool False if the item was not found, true otherwise.
      */
-    public function remove($datum)
+    public function remove(mixed $datum)
     {
         $found = false;
         $key   = null;
@@ -286,15 +284,11 @@ class PriorityQueue implements Countable, IteratorAggregate, Serializable
      */
     public function toArray($flag = self::EXTR_DATA)
     {
-        switch ($flag) {
-            case self::EXTR_BOTH:
-                return $this->items;
-            case self::EXTR_PRIORITY:
-                return array_map(static fn($item) => $item['priority'], $this->items);
-            case self::EXTR_DATA:
-            default:
-                return array_map(static fn($item) => $item['data'], $this->items);
-        }
+        return match ($flag) {
+            self::EXTR_BOTH => $this->items,
+            self::EXTR_PRIORITY => array_map(static fn($item): int => $item['priority'], $this->items),
+            default => array_map(static fn($item): mixed => $item['data'], $this->items),
+        };
     }
 
     /**
@@ -359,11 +353,11 @@ class PriorityQueue implements Countable, IteratorAggregate, Serializable
             $queue = new $this->queueClass();
             /** @psalm-var \SplPriorityQueue<TPriority, TValue> $queue */
             $this->queue = $queue;
-            /** @psalm-suppress DocblockTypeContradiction, MixedArgument */
+            /** @psalm-suppress DocblockTypeContradiction */
             if (! $this->queue instanceof \SplPriorityQueue) {
                 throw new Exception\DomainException(sprintf(
                     'PriorityQueue expects an internal queue of type SplPriorityQueue; received "%s"',
-                    get_class($this->queue)
+                    $queue::class
                 ));
             }
         }
