@@ -2,7 +2,7 @@
 
 // autoload_real.php @generated by Composer
 
-class ComposerAutoloaderInitcb14b996e01f21057472f2e6dc5c8cf9
+class ComposerAutoloaderInit829065e31ba829e6ee711b9f1f2c7cdb
 {
     private static $loader;
 
@@ -24,15 +24,15 @@ class ComposerAutoloaderInitcb14b996e01f21057472f2e6dc5c8cf9
 
         require __DIR__ . '/platform_check.php';
 
-        spl_autoload_register(array('ComposerAutoloaderInitcb14b996e01f21057472f2e6dc5c8cf9', 'loadClassLoader'), true, true);
+        spl_autoload_register(array('ComposerAutoloaderInit829065e31ba829e6ee711b9f1f2c7cdb', 'loadClassLoader'), true, true);
         self::$loader = $loader = new \Composer\Autoload\ClassLoader(\dirname(\dirname(__FILE__)));
-        spl_autoload_unregister(array('ComposerAutoloaderInitcb14b996e01f21057472f2e6dc5c8cf9', 'loadClassLoader'));
+        spl_autoload_unregister(array('ComposerAutoloaderInit829065e31ba829e6ee711b9f1f2c7cdb', 'loadClassLoader'));
 
         $useStaticLoader = PHP_VERSION_ID >= 50600 && !defined('HHVM_VERSION') && (!function_exists('zend_loader_file_encoded') || !zend_loader_file_encoded());
         if ($useStaticLoader) {
             require __DIR__ . '/autoload_static.php';
 
-            call_user_func(\Composer\Autoload\ComposerStaticInitcb14b996e01f21057472f2e6dc5c8cf9::getInitializer($loader));
+            call_user_func(\Composer\Autoload\ComposerStaticInit829065e31ba829e6ee711b9f1f2c7cdb::getInitializer($loader));
         } else {
             $map = require __DIR__ . '/autoload_namespaces.php';
             foreach ($map as $namespace => $path) {
