@@ -201,9 +201,11 @@ class Result implements Iterator, ResultInterface
                 'This result is a forward only result set, calling rewind() after moving forward is not supported'
             );
         }
-        $this->currentData     = $this->resource->fetch($this->fetchMode);
-        $this->currentComplete = true;
-        $this->position        = 0;
+        if (! $this->currentComplete) {
+            $this->currentData     = $this->resource->fetch($this->fetchMode);
+            $this->currentComplete = true;
+        }
+        $this->position = 0;
     }
 
     /**
