@@ -24,6 +24,6 @@ class Sql92 extends AbstractPlatform
             'Attempting to quote a value without specific driver level support'
             . ' can introduce security vulnerabilities in a production environment.'
         );
-        return '\'' . addcslashes($value, "\x00\n\r\\'\"\x1a") . '\'';
+        return '\'' . addcslashes($value ?? '', "\x00\n\r\\'\"\x1a") . '\'';
     }
 }
