@@ -6,6 +6,7 @@ use Closure;
 use Laminas\Db\Adapter\Driver\DriverInterface;
 use Laminas\Db\Adapter\ParameterContainer;
 use Laminas\Db\Adapter\Platform\PlatformInterface;
+use Laminas\Db\Sql\Predicate\PredicateInterface;
 
 use function array_key_exists;
 use function sprintf;
@@ -86,7 +87,7 @@ class Delete extends AbstractPreparableSql
     /**
      * Create where clause
      *
-     * @param Where|Closure|string|array $predicate
+     * @param Where|Closure|string|array|PredicateInterface $predicate
      * @param  string $combination One of the OP_* constants from Predicate\PredicateSet
      * @return $this Provides a fluent interface
      */
