@@ -7,6 +7,7 @@ namespace Laminas\Stdlib;
 use Traversable;
 
 use function array_shift;
+use function get_object_vars;
 use function is_array;
 use function is_callable;
 use function method_exists;
@@ -16,6 +17,10 @@ use function str_replace;
 use function strtolower;
 use function ucwords;
 
+/**
+ * @template TValue
+ * @implements ParameterObjectInterface<string, TValue>
+ */
 abstract class AbstractOptions implements ParameterObjectInterface
 {
     // phpcs:disable PSR2.Classes.PropertyDeclaration.Underscore,WebimpressCodingStandard.NamingConventions.ValidVariableName.NotCamelCapsProperty
@@ -33,7 +38,7 @@ abstract class AbstractOptions implements ParameterObjectInterface
     /**
      * Constructor
      *
-     * @param  array|Traversable|null $options
+     * @param  iterable<string, TValue>|AbstractOptions<TValue>|null $options
      */
     public function __construct($options = null)
     {
@@ -45,7 +50,7 @@ abstract class AbstractOptions implements ParameterObjectInterface
     /**
      * Set one or more configuration properties
      *
-     * @param  array|Traversable|AbstractOptions $options
+     * @param  iterable<string, TValue>|AbstractOptions<TValue> $options
      * @throws Exception\InvalidArgumentException
      * @return AbstractOptions Provides fluent interface
      */
@@ -77,19 +82,20 @@ abstract class AbstractOptions implements ParameterObjectInterface
     /**
      * Cast to array
      *
-     * @return array
+     * @return array<string, TValue>
      */
     public function toArray()
     {
         $array = [];
 
-        /** @param string[] $letters */
         $transform = static function (array $letters): string {
+            /** @var list<string> $letters */
             $letter = array_shift($letters);
             return '_' . strtolower($letter);
         };
 
-        foreach ($this as $key => $value) {
+        /** @psalm-var TValue $value */
+        foreach (get_object_vars($this) as $key => $value) {
             if ($key === '__strictMode__') {
                 continue;
             }
@@ -106,7 +112,7 @@ abstract class AbstractOptions implements ParameterObjectInterface
      * @see ParameterObject::__set()
      *
      * @param string $key
-     * @param mixed $value
+     * @param TValue|null $value
      * @throws Exception\BadMethodCallException
      * @return void
      */
@@ -137,7 +143,7 @@ abstract class AbstractOptions implements ParameterObjectInterface
      *
      * @param string $key
      * @throws Exception\BadMethodCallException
-     * @return mixed
+     * @return TValue
      */
     public function __get($key)
     {
