@@ -12,6 +12,10 @@ use function array_reverse;
 
 /**
  * ArrayObject that acts as a stack with regards to iteration
+ *
+ * @template TKey of array-key
+ * @template TValue
+ * @template-extends PhpArrayObject<TKey, TValue>
  */
 class ArrayStack extends PhpArrayObject
 {
@@ -21,7 +25,7 @@ class ArrayStack extends PhpArrayObject
      * Retrieve an array copy of the object, reverse its order, and return an
      * ArrayIterator with that reversed array.
      *
-     * @return ArrayIterator
+     * @return ArrayIterator<TKey, TValue>
      */
     #[ReturnTypeWillChange]
     public function getIterator()
