@@ -7,10 +7,8 @@ namespace Laminas\Stdlib;
 use Traversable;
 
 use function array_key_exists;
-use function get_class;
-use function gettype;
+use function get_debug_type;
 use function is_array;
-use function is_object;
 use function is_scalar;
 use function sprintf;
 
@@ -42,7 +40,7 @@ class Message implements MessageInterface
         if (! is_array($spec) && ! $spec instanceof Traversable) {
             throw new Exception\InvalidArgumentException(sprintf(
                 'Expected a string, array, or Traversable argument in first position; received "%s"',
-                is_object($spec) ? get_class($spec) : gettype($spec)
+                get_debug_type($spec)
             ));
         }
         foreach ($spec as $key => $value) {
