@@ -9,17 +9,22 @@ use Countable;
 use Serializable;
 use Traversable;
 
-/*
+/**
  * Basically, an ArrayObject. You could simply define something like:
  *     class QueryParams extends ArrayObject implements Parameters {}
  * and have 90% of the functionality
+ *
+ * @template TKey
+ * @template TValue
+ * @template-extends ArrayAccess<TKey, TValue>
+ * @template-extends Traversable<TKey, TValue>
  */
 interface ParametersInterface extends ArrayAccess, Countable, Serializable, Traversable
 {
     /**
      * Constructor
      *
-     * @param array $values
+     * @param array<TKey, TValue>|null $values
      */
     public function __construct(?array $values = null);
 
@@ -28,7 +33,7 @@ interface ParametersInterface extends ArrayAccess, Countable, Serializable, Trav
      *
      * Allow deserialization from standard array
      *
-     * @param array $values
+     * @param array<TKey, TValue> $values
      * @return mixed
      */
     public function fromArray(array $values);
@@ -48,7 +53,7 @@ interface ParametersInterface extends ArrayAccess, Countable, Serializable, Trav
      *
      * Allow serialization back to standard array
      *
-     * @return mixed
+     * @return array<TKey, TValue>
      */
     public function toArray();
 
@@ -57,24 +62,20 @@ interface ParametersInterface extends ArrayAccess, Countable, Serializable, Trav
      *
      * Allow serialization to query format; e.g., for PUT or POST requests
      *
-     * @return mixed
+     * @return string
      */
     public function toString();
 
     /**
-     * Get
-     *
-     * @param string $name
-     * @param mixed|null $default
+     * @param TKey $name
+     * @param TValue|null $default
      * @return mixed
      */
     public function get($name, $default = null);
 
     /**
-     * Set
-     *
-     * @param string $name
-     * @param mixed $value
+     * @param TKey $name
+     * @param TValue $value
      * @return ParametersInterface
      */
     public function set($name, $value);
