@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Laminas\Stdlib;
 
+use ReturnTypeWillChange;
 use Serializable;
 use UnexpectedValueException;
 
 use function array_key_exists;
-use function get_class;
-use function gettype;
+use function get_debug_type;
 use function is_array;
-use function is_object;
 use function serialize;
 use function sprintf;
 use function unserialize;
@@ -39,17 +38,18 @@ class SplPriorityQueue extends \SplPriorityQueue implements Serializable
      * Utilizes {@var $serial} to ensure that values of equal priority are
      * emitted in the same order in which they are inserted.
      *
-     * @param  TValue    $datum
+     * @param  TValue    $value
      * @param  TPriority $priority
      * @return void
      */
-    public function insert($datum, $priority)
+    #[ReturnTypeWillChange] // Inherited return type should be bool
+    public function insert($value, $priority)
     {
         if (! is_array($priority)) {
             $priority = [$priority, $this->serial--];
         }
 
-        parent::insert($datum, $priority);
+        parent::insert($value, $priority);
     }
 
     /**
@@ -129,7 +129,7 @@ class SplPriorityQueue extends \SplPriorityQueue implements Serializable
                 throw new UnexpectedValueException(sprintf(
                     'Cannot deserialize %s instance: corrupt item; expected array, received %s',
                     self::class,
-                    is_object($item) ? get_class($item) : gettype($item)
+                    get_debug_type($item)
                 ));
             }
 
